// src/web/src/api/schema.d.ts
export interface components {
    schemas: {
        /** Calorie */
        Calorie: {
            /** Amount */
            amount: number;
            /**
             * Unit
             * @constant
             */
            unit: "kcal";
        };
        /** DateRange */
        DateRange: {
            /**
             * From
             * Format: date
             */
            from: string;
            /**
             * To
             * Format: date
             */
            to: string;
        };
        /** FieldError */
        FieldError: {
            /** Field */
            field: string;
            /** Message */
            message: string;
        };
        /** IngredientOrigin */
        IngredientOrigin: {
            /** Ingredient */
            ingredient: string;
            /** Origin */
            origin: string;
        };
        /** Meal */
        Meal: {
            /**
             * Date
             * Format: date
             */
            date: string;
            /** Dishes */
            dishes: string[];
            calorie: components["schemas"]["Calorie"] | null;
            /** Nutrition */
            nutrition: components["schemas"]["Nutrient"][];
            /** Origininfo */
            originInfo: components["schemas"]["IngredientOrigin"][];
            /** Servingcount */
            servingCount: number | null;
        };
        /** MealSearchResponse */
        MealSearchResponse: {
            school: components["schemas"]["SelectedSchool"];
            range: components["schemas"]["DateRange"];
            /**
             * Mealtype
             * @constant
             */
            mealType: "LUNCH";
            /** Items */
            items: components["schemas"]["Meal"][];
        };
        /** Nutrient */
        Nutrient: {
            /** Name */
            name: string;
            /** Amount */
            amount: number;
            /** Unit */
            unit: string;
        };
        /** Pagination */
        Pagination: {
            /** Page */
            page: number;
            /** Pagesize */
            pageSize: number;
            /** Totalcount */
            totalCount: number;
        };
        /** ProblemDetails */
        ProblemDetails: {
            /**
             * Type
             * Format: uri
             */
            type: string;
            /** Title */
            title: string;
            /** Status */
            status: number;
            /** Detail */
            detail: string;
            /** Instance */
            instance: string;
            /** Code */
            code: string;
            /** Errors */
            errors?: components["schemas"]["FieldError"][] | null;
            /** Traceid */
            traceId: string;
        };
        /** SchoolSearchResponse */
        SchoolSearchResponse: {
            /** Items */
            items: components["schemas"]["SchoolSummary"][];
            pagination: components["schemas"]["Pagination"];
        };
        /** SchoolSummary */
        SchoolSummary: {
            /** Educationofficecode */
            educationOfficeCode: string;
            /** Educationofficename */
            educationOfficeName: string;
            /** Schoolcode */
            schoolCode: string;
            /** Name */
            name: string;
            /** Schooltype */
            schoolType: string;
            /** Region */
            region: string;
            /** Address */
            address: string | null;
        };
        /** SelectedSchool */
        SelectedSchool: {
            /** Educationofficecode */
            educationOfficeCode: string;
            /** Schoolcode */
            schoolCode: string;
            /** Name */
            name: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}